requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
APScheduler==3.10.4
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
pillow==10.4.0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
table2ascii==1.1.3
tzlocal==5.2
urllib3==2.2.2
wcwidth==0.2.13
yarl==1.9.11
